+++ src/werkzeug/datastructures/mixins.py
@@ -20,8 +20,12 @@ def _immutable_error(self: t.Any) -> t.NoReturn:
     raise TypeError(f"{type(self).__name__!r} objects are immutable")
 
 
-class ImmutableListMixin:
+class _ImmutableListMixin:
     """Makes a :class:`list` immutable.
+
+    .. deprecated:: 3.2
+        Will be removed in Werkzeug 3.3. Use ``collections.abc.Sequence``
+        instead.
 
     .. versionadded:: 0.5
 
@@ -73,8 +77,12 @@ class ImmutableListMixin:
         _immutable_error(self)
 
 
-class ImmutableDictMixin(t.Generic[K, V]):
+class _ImmutableDictMixin(t.Generic[K, V]):
     """Makes a :class:`dict` immutable.
+
+    .. deprecated:: 3.2
+        Will be removed in Werkzeug 3.3. Use ``collections.abc.Mapping``
+        instead.
 
     .. versionchanged:: 3.1
         Disallow ``|=`` operator.
@@ -90,14 +98,16 @@ class ImmutableDictMixin(t.Generic[K, V]):
     @t.overload
     def fromkeys(
         cls, keys: cabc.Iterable[K], value: None
-    ) -> ImmutableDictMixin[K, t.Any | None]: ...
+    ) -> _ImmutableDictMixin[K, t.Any | None]: ...
     @classmethod
     @t.overload
-    def fromkeys(cls, keys: cabc.Iterable[K], value: V) -> ImmutableDictMixin[K, V]: ...
+    def fromkeys(
+        cls, keys: cabc.Iterable[K], value: V
+    ) -> _ImmutableDictMixin[K, V]: ...
     @classmethod
     def fromkeys(
         cls, keys: cabc.Iterable[K], value: V | None = None
-    ) -> ImmutableDictMixin[K, t.Any | None] | ImmutableDictMixin[K, V]:
+    ) -> _ImmutableDictMixin[K, t.Any | None] | _ImmutableDictMixin[K, V]:
         instance = super().__new__(cls)
         instance.__init__(zip(keys, repeat(value)))  # type: ignore[misc]
         return instance
@@ -139,7 +149,7 @@ class ImmutableDictMixin(t.Generic[K, V]):
         _immutable_error(self)
 
 
-class ImmutableMultiDictMixin(ImmutableDictMixin[K, V]):
+class ImmutableMultiDictMixin(_ImmutableDictMixin[K, V]):
     """Makes a :class:`MultiDict` immutable.
 
     .. versionadded:: 0.5
@@ -230,7 +240,7 @@ class ImmutableHeadersMixin:
 
 def _always_update(f: F) -> F:
     def wrapper(
-        self: UpdateDictMixin[t.Any, t.Any], /, *args: t.Any, **kwargs: t.Any
+        self: _UpdateDictMixin[t.Any, t.Any], /, *args: t.Any, **kwargs: t.Any
     ) -> t.Any:
         rv = f(self, *args, **kwargs)
 
@@ -242,8 +252,11 @@ def _always_update(f: F) -> F:
     return update_wrapper(wrapper, f)  # type: ignore[return-value]
 
 
-class UpdateDictMixin(dict[K, V]):
+class _UpdateDictMixin(dict[K, V]):
     """Makes dicts call `self.on_update` on modifications.
+
+    .. deprecated:: 3.2
+        Will be removed in Werkzeug 3.3. Use ``CallbackDict`` instead.
 
     .. versionchanged:: 3.1
         Implement ``|=`` operator.
@@ -315,3 +328,26 @@ class UpdateDictMixin(dict[K, V]):
         self, other: cabc.Mapping[K, V] | cabc.Iterable[tuple[K, V]]
     ) -> te.Self:
         return super().__ior__(other)
+
+
+if not t.TYPE_CHECKING:
+
+    def __getattr__(name: str) -> t.Any:
+        alts = {
+            "ImmutableListMixin": "collections.abc.Sequence",
+            "ImmutableDictMixin": "collections.abc.Mapping",
+            "UpdateDictMixin": "CallbackDict",
+        }
+
+        if name in alts:
+            import warnings
+
+            warnings.warn(
+                f"The '{name}' class is deprecated and will be removed in"
+                f" Werkzeug 3.3. Use '{alts[name]}' instead.",
+                DeprecationWarning,
+                stacklevel=2,
+            )
+            return globals()[f"_{name}"]
+
+        raise AttributeError(name)