+++ tests/test_wrappers.py
@@ -110,7 +110,7 @@ def test_access_route():
         headers={"X-Forwarded-For": "192.168.1.2, 192.168.1.1"},
         environ_base={"REMOTE_ADDR": "192.168.1.3"},
     )
-    assert req.access_route == ["192.168.1.2", "192.168.1.1"]
+    assert req.access_route == ("192.168.1.2", "192.168.1.1")
     assert req.remote_addr == "192.168.1.3"
 
     req = wrappers.Request.from_values(environ_base={"REMOTE_ADDR": "192.168.1.3"})